requirements.txt
pandas
streamlit
deepeval
langchain
langchain_openai
langchain_chroma
python-dotenv